added mobile and desktop displays
